Proxy Builder + Multhitrheading
Proxy give the user implementation of proxy + saving the reusable information and when user makes operation more than once server don't need to make the operation twice,it  just returns the cachable value inside the proxy which saving efforts and time within backend.

If information is updated Multhithreading inside the proxy make parallel operation to update the information. This way cachable is updated and when  user requires the information it is already saved in the cash and server does not make anything except returning the cash inside the proxy